Refactor into packages.
