+++ src/main/java/hamiltonianguy/tasks/DeadlineTask.java
@@ -1,42 +1,47 @@
+package hamiltonianguy.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class EventTask extends Task {
-    private LocalDateTime from;
-    private LocalDateTime to;
+
+public class DeadlineTask extends Task {
+    private LocalDateTime by;
     private static final DateTimeFormatter INPUT_FORMAT = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
     private static final DateTimeFormatter OUTPUT_FORMAT = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mm a");
 
-    public EventTask(String desc, String from, String to) {
+    public DeadlineTask(String desc, String by) {
         super(desc);
         try {
-            this.from = LocalDateTime.parse(from, INPUT_FORMAT);
-            this.to = LocalDateTime.parse(to, INPUT_FORMAT);
+            this.by = LocalDateTime.parse(by, INPUT_FORMAT);
         } catch (DateTimeParseException e) {
             throw new IllegalArgumentException("OOPS!!! Invalid date format. Use yyyy-MM-dd HHmm (e.g., 2024-02-01 1800)");
         }
     }
 
-    public EventTask(String desc, LocalDateTime from, LocalDateTime to, Boolean isDone) {
-        super(desc, isDone);
-        this.from = from;
-        this.to = to;
+    public DeadlineTask(String desc, LocalDateTime by) {
+        super(desc);
+        this.by = by;
     }
+
+    public DeadlineTask(String desc, LocalDateTime by, boolean isDone) {
+        super(desc, isDone);
+        this.by = by;
+    }
+
 
     @Override
     public String getType() {
-        return "[E]";
+        return "[D]";
     }
 
     @Override
     public String toString() {
-        return getType() + super.toString() + " (from: " + from.format(OUTPUT_FORMAT) + " to: " + to.format(OUTPUT_FORMAT) + ")";
+        return getType() + super.toString() + " (by: " + by.format(OUTPUT_FORMAT) + ")";
     }
 
     @Override
     public String toFileString() {
-        return "E | " + (isDone() ? "1" : "0") + " | " + getDesc() + " | " + from.format(INPUT_FORMAT) + " | " + to.format(INPUT_FORMAT);
+        return "D | " + (isDone() ? "1" : "0") + " | " + getDesc() + " | " + by.format(INPUT_FORMAT);
     }
 }
-
